+++ purchase_allowed_product/models/product.py
@@ -11,9 +11,9 @@ class ProductProduct(models.Model):
     @api.model
     def search(self, args, offset=0, limit=None, order=None, count=False):
         restrict_supplier_id = self.env.context.get(
-            'restrict_supplier_id', False)
+            'restrict_supplier_id')
         use_only_supplied_product = self.env.context.get(
-            'use_only_supplied_product', False)
+            'use_only_supplied_product')
         if use_only_supplied_product:
             seller = self.env['res.partner'].browse(restrict_supplier_id)
             seller = seller.commercial_partner_id if seller.\
